+++ code/src/TransactionStartEvent.cpp
@@ -18,7 +18,12 @@ TransactionStartEvent::TransactionStartEvent() :
 {
 }
 
-TransactionStartEvent::TransactionStartEvent(ActorId ship_id, ActorId port_id, double ship_gold, double ship_rum, double port_rum, unsigned int rum_price)
+TransactionStartEvent::TransactionStartEvent(ActorId ship_id,
+					     ActorId port_id,
+					     unsigned int ship_gold,
+					     unsigned int ship_rum,
+					     unsigned int port_rum,
+					     unsigned int rum_price)
   : EventInterface(),
     ship_id(ship_id),
     port_id(port_id),
